@@ -14,43 +14,45 @@
   "dependencies": {
     "@angular-eslint/eslint-plugin": "^18.4.0",
     "@angular-eslint/schematics": "^18.4.0",
-    "@angular/animations": "^18.2.8",
-    "@angular/cdk": "^18.2.9",
-    "@angular/common": "^18.2.8",
-    "@angular/compiler": "^18.2.8",
-    "@angular/core": "^18.2.8",
-    "@angular/forms": "^18.2.8",
-    "@angular/material": "^18.2.9",
-    "@angular/platform-browser": "^18.2.8",
-    "@angular/platform-browser-dynamic": "^18.2.8",
-    "@angular/router": "^18.2.8",
-    "@apollo/client": "^3.0.0",
+    "@angular/animations": "^20.3.12",
+    "@angular/cdk": "^19.2.19",
+    "@angular/common": "^20.3.12",
+    "@angular/compiler": "^20.3.12",
+    "@angular/core": "^20.3.12",
+    "@angular/forms": "^20.3.12",
+    "@angular/material": "^19.2.19",
+    "@angular/platform-browser": "^20.3.12",
+    "@angular/platform-browser-dynamic": "^20.3.12",
+    "@angular/router": "^20.3.12",
+    "@apollo/client": "^4.0.9",
     "@geoman-io/leaflet-geoman-free": "^2.17.0",
-    "@ngrx/effects": "^18.1.1",
-    "@ngrx/store": "^18.1.1",
+    "@ngrx/effects": "^20.1.0",
+    "@ngrx/signals": "^20.1.0",
+    "@ngrx/store": "^20.1.0",
     "@stripe/stripe-js": "^5.2.0",
     "@tailwindcss/forms": "^0.5.9",
     "@types/geojson": "^7946.0.14",
     "@types/leaflet": "^1.9.12",
     "@types/leaflet-draw": "^1.0.11",
     "@types/leaflet.markercluster": "^1.5.4",
-    "apollo-angular": "^7.2.1",
-    "chart.js": "^4.4.5",
-    "graphql": "^16",
+    "apollo-angular": "^12.1.0",
+    "chart.js": "^4.5.1",
+    "class-validator": "^0.14.2",
+    "graphql": "^16.12.0",
     "jquery": "^3.7.1",
     "leaflet-draw": "^1.0.4",
     "leaflet-geosearch": "^4.0.0",
     "leaflet.markerclusterv2": "^2.0.1",
-    "ng2-charts": "^6.0.1",
-    "ngrx-store-localstorage": "^18.0.0",
+    "ng2-charts": "^8.0.0",
+    "ngrx-store-localstorage": "^20.0.0",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.14.10"
+    "zone.js": "~0.15.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^18.2.9",
-    "@angular/cli": "~18.2.9",
-    "@angular/compiler-cli": "^18.2.8",
+    "@angular/build": "^20.3.10",
+    "@angular/cli": "~20.3.10",
+    "@angular/compiler-cli": "^20.3.12",
     "@types/google.maps": "^3.58.1",
     "@types/jasmine": "~4.3.0",
     "@types/jquery": "^3.5.30",
@@ -69,7 +71,7 @@
     "prettier": "^3.3.3",
     "prettier-eslint": "^16.3.0",
     "tailwindcss": "^3.4.15",
-    "typescript": "~5.4.5",
+    "typescript": "~5.8.3",
     "typescript-eslint": "8.2.0"
   }
 }
