@@ -1,19 +1,21 @@
 
-import { Component } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { CommonModule } from '@angular/common';
+import { Component, inject } from '@angular/core';
+import { FormBuilder, FormGroup, FormsModule, ReactiveFormsModule, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
-import { Store } from '@ngrx/store';
-import { Observable } from 'rxjs';
 import { User } from 'src/app/Model/User';
 import { AuthentificationApi } from 'src/app/Services/AuthentificationApi.service';
 import { SeoService } from 'src/app/Services/Seo.service';
 import { Picture, RandomPictureService } from 'src/app/Shared/Services/RandomPicture.service';
-import { selectIsAuthenticated } from 'src/app/Store/Selectors/user.selector';
+import { UserStore } from 'src/app/Store/user/user.store';
 import { StrongPasswordRegx } from 'src/app/Utils/Const/StrongPasswordRegex';
 import { environment } from 'src/environments/environment';
 
 @Component({
-  templateUrl: './Authentification.view.html'
+    templateUrl: './Authentification.view.html',
+    standalone: true,
+    imports: [FormsModule, ReactiveFormsModule, CommonModule],
+    providers: [AuthentificationApi, Router, SeoService, RandomPictureService, FormBuilder, UserStore],
 })
 export class AuthenticationView {
 
@@ -24,28 +26,19 @@ export class AuthenticationView {
   registerForm: FormGroup;
   errorMessage = '';
   picture: Picture;
-  isAuthenticated$: Observable<boolean>;
 
-  constructor(private route: Router,
-    private seoService: SeoService, 
-    private randomPictureService: RandomPictureService,
-    private authentificationApi: AuthentificationApi,
-    private store: Store,
-    private fb: FormBuilder) { 
+  #authentificationApi = inject(AuthentificationApi);
+  #route = inject(Router);
+  #seoService = inject(SeoService);
+  #randomPictureService = inject(RandomPictureService);
+  #fb = inject(FormBuilder);
 
-    this.picture = this.randomPictureService.getPictureRandom();
+  constructor() { 
 
-    //Redirect if already connected
-    this.isAuthenticated$ = this.store.select(selectIsAuthenticated);
-    this.isAuthenticated$.subscribe((isAuth) => {
-      if(isAuth){
-        this.route.navigateByUrl('/dashboard');
-      }
-    })
+    this.picture = this.#randomPictureService.getPictureRandom();
 
-    this.seoService.generateTags("S'authentifier sur Disastream","Inscrivez-vous sur Disastream pour être notifiés des dernières catastrophes naturelles",`${this.s3BasePath}/background/avalanche.jpg`);
-
-    this.registerForm = this.fb.group({
+    this.#seoService.generateTags("S'authentifier sur Disastream","Inscrivez-vous sur Disastream pour être notifiés des dernières catastrophes naturelles",`${this.s3BasePath}/background/avalanche.jpg`);
+    this.registerForm = this.#fb.group({
       username: ['', [Validators.required, Validators.minLength(3), Validators.maxLength(50)]],
       mail: ['', [Validators.required, Validators.email]],
       password: ['', [Validators.required, Validators.pattern(StrongPasswordRegx)]],
@@ -55,7 +48,7 @@ export class AuthenticationView {
   }
 
   authenticate(): void {
-    this.authentificationApi.googleSignin();
+    this.#authentificationApi.googleSignin();
   }
 
   get passwordFormField() {
@@ -64,11 +57,10 @@ export class AuthenticationView {
 
   onSubmit(): void {
     if (this.registerForm.invalid) return;
-
-    this.authentificationApi.register(this.registerForm.value).subscribe({
+    this.#authentificationApi.register(this.registerForm.value).subscribe({
       next: (user: User) => {
         // redirection ou message de succès
-        this.route.navigateByUrl(`/?mail=${user.mail}`);
+        this.#route.navigateByUrl(`/?mail=${user.mail}`);
       },
       error: (err) => {
         this.errorMessage = err.error.message || 'Erreur inconnue';
