@@ -1,23 +1,35 @@
-import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { Component, inject } from '@angular/core';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { User } from 'src/app/Model/User';
 import { AuthentificationApi } from 'src/app/Services/AuthentificationApi.service';
 import { UserApiService } from 'src/app/Services/UserApiService';
+import { AuthStore } from 'src/app/Store/auth/auth.store';
+import { UserStore } from 'src/app/Store/user/user.store';
 
 @Component({
-  selector: 'app-user-profile',
-  templateUrl: './user-profile.component.html',
-  styleUrls: ['./user-profile.component.css']
+    selector: 'app-user-profile',
+    templateUrl: './user-profile.component.html',
+    styleUrls: ['./user-profile.component.css'],
+    standalone: true,
+    imports: [CommonModule, ReactiveFormsModule, FormsModule],
+    providers: [UserApiService, AuthentificationApi, UserStore]
 })
 export class UserProfileComponent {
 
   user: User = new User();
 
-  constructor(private readonly userService: UserApiService, private readonly authService: AuthentificationApi){
+  #authStore = inject(AuthStore);
+  #authService = inject(AuthentificationApi);
+  #userService = inject(UserApiService);
+  protected readonly userStore = inject(UserStore);
+
+  constructor(){
     this.getMyProfile();
   }
 
   getMyProfile(){
-    this.userService.getMyProfile().subscribe((user) => {
+    this.#userService.getMyProfile().subscribe((user) => {
       this.user = user;
     })
   }
@@ -35,6 +47,6 @@ export class UserProfileComponent {
   }
 
   logout(){
-    this.authService.logOut();
+    this.#authStore.logout();
   }
 }
