@@ -1,4 +1,8 @@
+import { IsEmail, IsStrongPassword } from "class-validator";
+
 export class UserLoginDto {
-    mail = '';
-    password = '';
+  @IsEmail()
+  mail: string;
+  @IsStrongPassword()
+  password: string;
 }
