@@ -1,14 +1,18 @@
 
+import { CommonModule } from '@angular/common';
 import { Component, EventEmitter, Output } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { FormBuilder, FormGroup, ReactiveFormsModule, Validators } from '@angular/forms';
 import { MailAlert } from 'src/app/Model/MailAlert';
 import { AlertApiService } from 'src/app/Services/AlertApiService';
 import { ToastrService } from 'src/app/Shared/Services/Toastr.service';
+import { SharedModule } from 'src/app/Shared/Shared.module';
 import { environment } from 'src/environments/environment';
 
 @Component({
     selector: "app-mail-alert-modal",
-    templateUrl: './AddMailAlert.modal.html'
+    templateUrl: './AddMailAlert.modal.html',
+    standalone: true,
+    imports: [SharedModule, CommonModule, ReactiveFormsModule],
 })
 export class AddMailAlert {
 
