@@ -1,3 +1,4 @@
+import { CommonModule } from "@angular/common";
 import { Component } from "@angular/core";
 import L from "leaflet";
 import { DetailService } from "src/app/Services/DetailService";
@@ -5,7 +6,9 @@ import { DetailService } from "src/app/Services/DetailService";
 @Component({
     selector: "app-disaster-detail",
     templateUrl: './disaster-detail.component.html',
-    styleUrls: ['./disaster-detail.component.css']
+    styleUrls: ['./disaster-detail.component.css'],
+    standalone: true,
+    imports: [CommonModule]
 })
 export class DisasterDetailComponent {
     disaster$ = this.detailService.disasterDetail$;
