@@ -1,5 +1,6 @@
+import { CommonModule } from '@angular/common';
 import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { FormBuilder, FormGroup, FormsModule, ReactiveFormsModule, Validators } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 import { AuthentificationApi } from 'src/app/Services/AuthentificationApi.service';
 import { ToastrService } from 'src/app/Shared/Services/Toastr.service';
@@ -7,7 +8,9 @@ import { StrongPasswordRegx } from 'src/app/Utils/Const/StrongPasswordRegex';
 import { environment } from 'src/environments/environment.prod';
 
 @Component({
-  templateUrl: './ChangePassword.view.html'
+    templateUrl: './ChangePassword.view.html',
+    standalone: true,
+    imports: [FormsModule, ReactiveFormsModule, CommonModule]
 })
 export class ChangePasswordView implements OnInit {
 
