@@ -1,13 +1,17 @@
 
+import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { FormBuilder, FormGroup, FormsModule, ReactiveFormsModule, Validators } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 import { AuthentificationApi } from 'src/app/Services/AuthentificationApi.service';
 import { ToastrService } from 'src/app/Shared/Services/Toastr.service';
+import { SharedModule } from 'src/app/Shared/Shared.module';
 import { environment } from 'src/environments/environment';
 
 @Component({
-  templateUrl: './ConfirmEmail.view.html'
+    templateUrl: './ConfirmEmail.view.html',
+    standalone: true,
+    imports: [FormsModule,ReactiveFormsModule, SharedModule, CommonModule]
 })
 export class ConfirmEmailView {
 
