@@ -4,19 +4,19 @@ import { AleaCategoryDto } from "src/app/DTO/AleaCategory.dto";
 import { Alea } from "src/app/Model/Alea";
 import { Alert } from "src/app/Model/Alert";
 import { PublicApiService } from "src/app/Services/PublicApi.service";
-import { ToastrService } from "src/app/Shared/Services/toastr.service";
+import { ToastrService } from "src/app/Shared/Services/Toastr.service";
 import { environment } from "src/environments/environment";
 
-class AleaVM {
+export class AleaVM {
     alea: Alea;
     selected = false;
 
-    constructor(alea: Alea){
+    constructor(alea: Alea) {
         this.alea = alea;
     }
 }
 
-class AleaCategoryVM {
+export class AleaCategoryVM {
     category = "";
     aleas: AleaVM[] = [];
 }
@@ -35,11 +35,11 @@ export class AleaTypesComponent implements OnInit, OnDestroy {
 
     private aleaTypeAlertSubscription!: Subscription;
     @Input() loadingAlert!: Observable<Alert>;
-    
+
     @Output() aleaChange = new EventEmitter<Alea[]>();
     @Output() completeStep = new EventEmitter<string>();
 
-    constructor(private readonly publicApiService: PublicApiService, private readonly toastrService: ToastrService){
+    constructor(private readonly publicApiService: PublicApiService, private readonly toastrService: ToastrService) {
         this.getAleas();
     }
 
@@ -48,7 +48,7 @@ export class AleaTypesComponent implements OnInit, OnDestroy {
             console.log(alert);
             this.selectedAleaTypes = alert.aleas;
             this.categories.forEach(item => item.aleas.forEach(aleaVM => {
-                if(this.selectedAleaTypes.find(alea => alea.name == aleaVM.alea.name)){
+                if (this.selectedAleaTypes.find(alea => alea.name == aleaVM.alea.name)) {
                     aleaVM.selected = true;
                 }
             }))
@@ -59,50 +59,50 @@ export class AleaTypesComponent implements OnInit, OnDestroy {
         this.aleaTypeAlertSubscription?.unsubscribe();
     }
 
-    getAleas(){
+    getAleas() {
         this.publicApiService.getAleasByCategory().subscribe((v) => {
             const aleasByCategory: AleaCategoryVM[] = [];
             v.forEach((item) => {
-              const aleaCategoryDto = item as AleaCategoryDto;
-              if(!aleasByCategory.find((cat) => aleaCategoryDto.category_name == (cat as AleaCategoryVM).category)){
-                aleasByCategory.push({
-                      category: aleaCategoryDto.category_name,
-                      aleas: [{
+                const aleaCategoryDto = item as AleaCategoryDto;
+                if (!aleasByCategory.find((cat) => aleaCategoryDto.category_name == (cat as AleaCategoryVM).category)) {
+                    aleasByCategory.push({
+                        category: aleaCategoryDto.category_name,
+                        aleas: [{
+                            alea: {
+                                id: aleaCategoryDto.alea_id,
+                                name: aleaCategoryDto.alea_name,
+                                label: aleaCategoryDto.alea_label,
+                            },
+                            selected: false,
+                        }]
+                    })
+                }
+                else {
+                    aleasByCategory.find((cat) => (cat as AleaCategoryVM).category == aleaCategoryDto.category_name)?.aleas.push({
                         alea: {
                             id: aleaCategoryDto.alea_id,
                             name: aleaCategoryDto.alea_name,
                             label: aleaCategoryDto.alea_label,
                         },
                         selected: false,
-                      }]
-                  })
-              }
-              else{
-                aleasByCategory.find((cat) => (cat as AleaCategoryVM).category == aleaCategoryDto.category_name)?.aleas.push({
-                  alea: {
-                  id: aleaCategoryDto.alea_id,
-                  name: aleaCategoryDto.alea_name,
-                  label: aleaCategoryDto.alea_label,
-                },
-                selected: false,
-                });
-              }
+                    });
+                }
             })
             this.categories = aleasByCategory;
         })
     }
 
-    complete(){
+    complete() {
         this.selectedAleaTypes = [];
         this.categories.forEach(item => item.aleas.forEach(aleaVM => {
-            if(aleaVM.selected){
+            if (aleaVM.selected) {
                 this.selectedAleaTypes.push(aleaVM.alea);
             }
         }))
         this.aleaChange.emit(this.selectedAleaTypes);
     }
 
-    selectAlea(alea: AleaVM){
+    selectAlea(alea: AleaVM) {
         alea.selected = !alea.selected;
         this.complete();
     }
