@@ -1,5 +1,5 @@
 import { ChangeDetectionStrategy, ChangeDetectorRef, Component, HostListener, ViewChild, inject } from "@angular/core";
-import { FormBuilder, Validators } from "@angular/forms";
+import { FormBuilder, ReactiveFormsModule, Validators } from "@angular/forms";
 import { ActivatedRoute, Router } from "@angular/router";
 import { FeatureCollection } from "geojson";
 import L from "leaflet";
@@ -17,420 +17,427 @@ import { GeographyApiService } from "src/app/Services/GeographyApi.service";
 import { ShapeService } from "src/app/Map/Services/shape.service";
 import { EndAlertComponent } from "src/app/Modals/EndAlert/EndAlert.modal";
 import { environment } from "src/environments/environment";
+import { CommonModule } from "@angular/common";
+import { SearchPlace } from "src/app/Modals/SearchPlace/SearchPlace.modal";
+import { MapComponent } from "src/app/Map/Components/map/map.component";
 
-class AleaVM {
+export class AleaVM {
   alea: Alea;
   selected = false;
 
-  constructor(alea: Alea){
-      this.alea = alea;
+  constructor(alea: Alea) {
+    this.alea = alea;
   }
 }
 
-class AleaCategoryVM {
+export class AleaCategoryVM {
   category = "";
   aleas: AleaVM[] = [];
 }
 
 @Component({
-    templateUrl: './NewAlert.component.html',
-    changeDetection: ChangeDetectionStrategy.OnPush
+  templateUrl: './NewAlert.component.html',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: true,
+  imports: [AddMailAlert, EndAlertComponent, CommonModule, ReactiveFormsModule, SearchPlace, MapComponent],
 })
 export class NewAlertView {
-  
-    #env = environment;
-    protected s3BasePath = this.#env.settings.s3_bucket;
-    private _formBuilder = inject(FormBuilder);
 
-    public alert: Alert = new Alert();
-    public loadedAlert: Subject<Alert> = new Subject<Alert>();
-    panelVisible = true;
-    editShape = false;
+  #env = environment;
+  protected s3BasePath = this.#env.settings.s3_bucket;
+  private _formBuilder = inject(FormBuilder);
 
-    public categories: AleaCategoryVM[] = [];
-    public selectedAleaTypes: Alea[] = [];
+  public alert: Alert = new Alert();
+  public loadedAlert: Subject<Alert> = new Subject<Alert>();
+  panelVisible = true;
+  editShape = false;
 
-    //Change detector to update component manually
-    private cd = inject(ChangeDetectorRef)
+  public categories: AleaCategoryVM[] = [];
+  public selectedAleaTypes: Alea[] = [];
 
-    editMode = false;
+  //Change detector to update component manually
+  private cd = inject(ChangeDetectorRef)
 
-    //Area
-    areaMap?: L.Map;
-    actualLayer?: L.GeoJSON;
-    countriesLayer?: L.GeoJSON = new L.GeoJSON(null);
-    public locationBox?: L.LatLngBounds;
+  editMode = false;
 
-    //team
-    public mailAlerts: MailAlert[] = [];
-    selectedMailIds: number[] = [];
-    
-    @ViewChild('mailAlertModal') mailAlertModal?: AddMailAlert;
-    @ViewChild('endAlertModal') endAlertModal?: EndAlertComponent;
+  //Area
+  areaMap?: L.Map;
+  actualLayer?: L.GeoJSON;
+  countriesLayer?: L.GeoJSON = new L.GeoJSON(null);
+  public locationBox?: L.LatLngBounds;
 
-    formGroup = this._formBuilder.group({
-      name: ['', Validators.required],
-    });
+  //team
+  public mailAlerts: MailAlert[] = [];
+  selectedMailIds: number[] = [];
 
-    constructor(private alertApiService: AlertApiService, 
-      private router: Router, 
-      private route: ActivatedRoute, 
-      private toastrService: ToastrService,
-      private publicApiService: PublicApiService,
-      private shapeService: ShapeService,
-      private geographyService: GeographyApiService){
+  @ViewChild('mailAlertModal') mailAlertModal?: AddMailAlert;
+  @ViewChild('endAlertModal') endAlertModal?: EndAlertComponent;
 
-      this.getAleas();
-      this.getTeamMembersMail();
+  formGroup = this._formBuilder.group({
+    name: ['', Validators.required],
+  });
 
-      //Edit alert
-      if(this.route.snapshot.queryParamMap.get('id') != null){
-        this.editMode = true;
-        const id = parseInt(this.route.snapshot.queryParamMap.get('id')!);
-        this.alertApiService.getAlertById(id).subscribe({
-          next: (alert) => {
+  constructor(private alertApiService: AlertApiService,
+    private router: Router,
+    private route: ActivatedRoute,
+    private toastrService: ToastrService,
+    private publicApiService: PublicApiService,
+    private shapeService: ShapeService,
+    private geographyService: GeographyApiService) {
 
-            this.alert = alert;
+    this.getAleas();
+    this.getTeamMembersMail();
 
-            //feed area
-            const geo = new L.GeoJSON(alert.areas);
-            this.actualLayer?.addLayer(geo);
-            this.addShapeToMap();
+    //Edit alert
+    if (this.route.snapshot.queryParamMap.get('id') != null) {
+      this.editMode = true;
+      const id = parseInt(this.route.snapshot.queryParamMap.get('id')!);
+      this.alertApiService.getAlertById(id).subscribe({
+        next: (alert) => {
 
-            //feed alea selected
-            this.categories.forEach(item => item.aleas.forEach(aleaVM => {
-              if(this.alert.aleas.find(alea => alea.name == aleaVM.alea.name)){
-                  aleaVM.selected = true;
-              }
-            }))
+          this.alert = alert;
 
-            //feed team users
-            this.selectedMailIds = (this.alert.mailAlerts) ? this.alert.mailAlerts.map(x => x.id) : [];
-            this.updateComponent();
-          },
-          error: (error) => { 
-            if(error.status == 403){
-              this.router.navigateByUrl('dashboard/alerts/manage').then(() => { this.toastrService.error('Vous n\'êtes pas autorisé à accéder à cette alerte'); 
-            })}
-            else {
-              this.toastrService.error(error?.message);
-            }
-          }
-        })
-      }
-      
-    }
-
-    /**
-     * Show side panel to make alert
-     */
-    showPanel(){
-      this.panelVisible = !this.panelVisible;
-      if(this.panelVisible && this.areaMap != null){
-        this.areaMap!.pm.disableGlobalEditMode();
-      }
-    }
-
-    /*********************AREA*************************/
-
-    /**
-     * Receive map from the map component
-     * @param map 
-     */
-    receiveMap(map: L.Map) {
-      this.areaMap = map;
-    }
-    
-    /**
-     * Receive layer from the map component
-     * @param layer 
-     */
-    receiveLayer(layer: L.LayerGroup) {
-        this.actualLayer = new L.GeoJSON(null,{
-            style: {
-                fillColor: '#6a5ac7'
-            }
-        });
-        if(this.areaMap !== undefined){
-            this.actualLayer!.addTo(this.areaMap);
-            this.actualLayer?.bringToFront()
-            this.areaMap.pm.setGlobalOptions({
-                layerGroup: this.actualLayer,
-            });
-            this.areaMap.on('pm:create', this.addShapeToMap, this);
-        }
-    }
-
-    /**
-     * Hide countries and show panel after country selection
-     */
-    hideCountries(){
-      this.countriesLayer?.removeFrom(this.areaMap!);
-      if(!this.panelVisible){
-        this.showPanel();
-      }
-    }
-
-    /**
-     * Show countries on map
-     */
-    showCountries(){
-      if(this.countriesLayer?.getLayers().length == 0){
-        this.getCountries();
-      }
-      else{
-        this.countriesLayer?.addTo(this.areaMap!);
-      }
-      if(this.panelVisible){
-        this.showPanel();
-      }
-    }
-
-    /**
-     * Get countries and add them on map
-     */
-    getCountries(){
-      this.geographyService.getCountries().subscribe(countries => {
-        countries.forEach(country => {
-          this.shapeService.makeCountriesShape(this.areaMap!,this.countriesLayer!,country, this.clickOnCountry, this);
-        })
-      });
-    }
-
-    /**
-     * Country selection
-     * @param e 
-     */
-    clickOnCountry(e){
-      const geometry = (e.sourceTarget?.feature?.geometry);
-      const idCountry = (e.sourceTarget?.feature?.properties?.id);
-      this.deleteArea();
-      if(this.actualLayer != null && geometry != null){
-          const geo = new L.GeoJSON(geometry);
+          //feed area
+          const geo = new L.GeoJSON(alert.areas);
           this.actualLayer?.addLayer(geo);
-          this.actualLayer?.addTo(this.areaMap!)
-          const collection = this.actualLayer?.toGeoJSON() as FeatureCollection;
-          this.alert.areas = collection?.features[0]?.geometry;
-          this.alert.countryId = idCountry;
-          this.alert.isCountryShape = true;
-      }
-      this.hideCountries();
-    }
+          this.addShapeToMap();
 
-    /**
-     * Quit edit mode with esc key
-     */
-    @HostListener('keydown.esc', ['$event'])
-    onEsc() {
-      if(this.areaMap != null){
-        this.areaMap.pm.disableDraw();
-        if(!this.panelVisible){
-          this.showPanel();
-        }
-      }
-    }
+          //feed alea selected
+          this.categories.forEach(item => item.aleas.forEach(aleaVM => {
+            if (this.alert.aleas.find(alea => alea.name == aleaVM.alea.name)) {
+              aleaVM.selected = true;
+            }
+          }))
 
-    /**
-     * Draw polygon form by clicking the button
-     */
-    drawPolygon(): void {
-      this.hideCountries()
-      if(this.areaMap !== undefined){
-          if(this.panelVisible){
-            this.showPanel();
+          //feed team users
+          this.selectedMailIds = (this.alert.mailAlerts) ? this.alert.mailAlerts.map(x => x.id) : [];
+          this.updateComponent();
+        },
+        error: (error) => {
+          if (error.status == 403) {
+            this.router.navigateByUrl('dashboard/alerts/manage').then(() => {
+              this.toastrService.error('Vous n\'êtes pas autorisé à accéder à cette alerte');
+            })
           }
-          // new L.Draw.Polygon(this.areaMap as (L.DrawMap), { shapeOptions: { stroke: true, fillColor: '#6a5ac7' }}).enable();
-          this.areaMap.pm.enableDraw("Polygon", {
-              snappable: true,
-              snapDistance: 5,
-          });
+          else {
+            this.toastrService.error(error?.message);
+          }
+        }
+      })
+    }
+
+  }
+
+  /**
+   * Show side panel to make alert
+   */
+  showPanel() {
+    this.panelVisible = !this.panelVisible;
+    if (this.panelVisible && this.areaMap != null) {
+      this.areaMap!.pm.disableGlobalEditMode();
+    }
+  }
+
+  /*********************AREA*************************/
+
+  /**
+   * Receive map from the map component
+   * @param map 
+   */
+  receiveMap(map: L.Map) {
+    this.areaMap = map;
+  }
+
+  /**
+   * Receive layer from the map component
+   * @param layer 
+   */
+  receiveLayer(layer: L.LayerGroup) {
+    this.actualLayer = new L.GeoJSON(null, {
+      style: {
+        fillColor: '#6a5ac7'
+      }
+    });
+    if (this.areaMap !== undefined) {
+      this.actualLayer!.addTo(this.areaMap);
+      this.actualLayer?.bringToFront()
+      this.areaMap.pm.setGlobalOptions({
+        layerGroup: this.actualLayer,
+      });
+      this.areaMap.on('pm:create', this.addShapeToMap, this);
+    }
+  }
+
+  /**
+   * Hide countries and show panel after country selection
+   */
+  hideCountries() {
+    this.countriesLayer?.removeFrom(this.areaMap!);
+    if (!this.panelVisible) {
+      this.showPanel();
+    }
+  }
+
+  /**
+   * Show countries on map
+   */
+  showCountries() {
+    if (this.countriesLayer?.getLayers().length == 0) {
+      this.getCountries();
+    }
+    else {
+      this.countriesLayer?.addTo(this.areaMap!);
+    }
+    if (this.panelVisible) {
+      this.showPanel();
+    }
+  }
+
+  /**
+   * Get countries and add them on map
+   */
+  getCountries() {
+    this.geographyService.getCountries().subscribe(countries => {
+      countries.forEach(country => {
+        this.shapeService.makeCountriesShape(this.areaMap!, this.countriesLayer!, country, this.clickOnCountry, this);
+      })
+    });
+  }
+
+  /**
+   * Country selection
+   * @param e 
+   */
+  clickOnCountry(e) {
+    const geometry = (e.sourceTarget?.feature?.geometry);
+    const idCountry = (e.sourceTarget?.feature?.properties?.id);
+    this.deleteArea();
+    if (this.actualLayer != null && geometry != null) {
+      const geo = new L.GeoJSON(geometry);
+      this.actualLayer?.addLayer(geo);
+      this.actualLayer?.addTo(this.areaMap!)
+      const collection = this.actualLayer?.toGeoJSON() as FeatureCollection;
+      this.alert.areas = collection?.features[0]?.geometry;
+      this.alert.countryId = idCountry;
+      this.alert.isCountryShape = true;
+    }
+    this.hideCountries();
+  }
+
+  /**
+   * Quit edit mode with esc key
+   */
+  @HostListener('keydown.esc', ['$event'])
+  onEsc() {
+    if (this.areaMap != null) {
+      this.areaMap.pm.disableDraw();
+      if (!this.panelVisible) {
+        this.showPanel();
       }
     }
+  }
 
-    /**
-     * Delete area
-     */
-    deleteArea(){
-      this.actualLayer?.clearLayers();
-      this.alert.isCountryShape = false;
-      this.alert.areas = null;
-      this.alert.countryId = null;
-    }
-
-    /**
-     * Edit custom area
-     */
-    editArea(){
-      this.showPanel();
-      this.areaMap!.pm.enableGlobalEditMode({
+  /**
+   * Draw polygon form by clicking the button
+   */
+  drawPolygon(): void {
+    this.hideCountries()
+    if (this.areaMap !== undefined) {
+      if (this.panelVisible) {
+        this.showPanel();
+      }
+      // new L.Draw.Polygon(this.areaMap as (L.DrawMap), { shapeOptions: { stroke: true, fillColor: '#6a5ac7' }}).enable();
+      this.areaMap.pm.enableDraw("Polygon", {
         snappable: true,
-        snapDistance: 50
+        snapDistance: 5,
       });
     }
+  }
 
-    
-    /**
-     * After adding shape (polygon or circle) by button
-     * @param e 
-     */
-    addShapeToMap(){
+  /**
+   * Delete area
+   */
+  deleteArea() {
+    this.actualLayer?.clearLayers();
+    this.alert.isCountryShape = false;
+    this.alert.areas = null;
+    this.alert.countryId = null;
+  }
 
-      if(this.areaMap != null){
+  /**
+   * Edit custom area
+   */
+  editArea() {
+    this.showPanel();
+    this.areaMap!.pm.enableGlobalEditMode({
+      snappable: true,
+      snapDistance: 50
+    });
+  }
 
-        this.actualLayer?.setStyle({weight: 3,fillColor:'#ffffff', color:'white'});
-        this.actualLayer?.addTo(this.areaMap!);
 
-        // this.areaMap!.pm.disableDraw();
-        // this.areaMap!.pm.enableGlobalEditMode({
-        //     snappable: true,
-        //     snapDistance: 50
-        // });
+  /**
+   * After adding shape (polygon or circle) by button
+   * @param e 
+   */
+  addShapeToMap() {
 
-        if(this.actualLayer != null){
-          const collection = this.actualLayer?.toGeoJSON() as FeatureCollection;
-          this.alert.areas = collection?.features[0]?.geometry;
-        }
-  
-        if(!this.panelVisible){
-          this.showPanel();
-        }
+    if (this.areaMap != null) {
+
+      this.actualLayer?.setStyle({ weight: 3, fillColor: '#ffffff', color: 'white' });
+      this.actualLayer?.addTo(this.areaMap!);
+
+      // this.areaMap!.pm.disableDraw();
+      // this.areaMap!.pm.enableGlobalEditMode({
+      //     snappable: true,
+      //     snapDistance: 50
+      // });
+
+      if (this.actualLayer != null) {
+        const collection = this.actualLayer?.toGeoJSON() as FeatureCollection;
+        this.alert.areas = collection?.features[0]?.geometry;
+      }
+
+      if (!this.panelVisible) {
+        this.showPanel();
       }
     }
+  }
 
-    /*********************ALEAS*************************/
+  /*********************ALEAS*************************/
 
-    /**
-     * Select alea to monitor
-     * @param alea 
-     */
-    selectAlea(alea: AleaVM){
-      alea.selected = !alea.selected;
-      this.alert.aleas = [];
-      this.categories.forEach(item => item.aleas.forEach(aleaVM => {
-          if(aleaVM.selected){
-              this.alert.aleas.push(aleaVM.alea);
-          }
-      }))
-    }
+  /**
+   * Select alea to monitor
+   * @param alea 
+   */
+  selectAlea(alea: AleaVM) {
+    alea.selected = !alea.selected;
+    this.alert.aleas = [];
+    this.categories.forEach(item => item.aleas.forEach(aleaVM => {
+      if (aleaVM.selected) {
+        this.alert.aleas.push(aleaVM.alea);
+      }
+    }))
+  }
 
-    /**
-     * 
-     */
-    getAleas(){
-      this.publicApiService.getAleasByCategory().subscribe((v) => {
-          const aleasByCategory: AleaCategoryVM[] = [];
-          v.forEach((item) => {
-            const aleaCategoryDto = item as AleaCategoryDto;
-            if(!aleasByCategory.find((cat) => aleaCategoryDto.category_name == (cat as AleaCategoryVM).category)){
-              aleasByCategory.push({
-                    category: aleaCategoryDto.category_name,
-                    aleas: [{
-                      alea: {
-                          id: aleaCategoryDto.alea_id,
-                          name: aleaCategoryDto.alea_name,
-                          label: aleaCategoryDto.alea_label,
-                      },
-                      selected: false,
-                    }]
-                })
-            }
-            else{
-              aleasByCategory.find((cat) => (cat as AleaCategoryVM).category == aleaCategoryDto.category_name)?.aleas.push({
-                alea: {
+  /**
+   * 
+   */
+  getAleas() {
+    this.publicApiService.getAleasByCategory().subscribe((v) => {
+      const aleasByCategory: AleaCategoryVM[] = [];
+      v.forEach((item) => {
+        const aleaCategoryDto = item as AleaCategoryDto;
+        if (!aleasByCategory.find((cat) => aleaCategoryDto.category_name == (cat as AleaCategoryVM).category)) {
+          aleasByCategory.push({
+            category: aleaCategoryDto.category_name,
+            aleas: [{
+              alea: {
                 id: aleaCategoryDto.alea_id,
                 name: aleaCategoryDto.alea_name,
                 label: aleaCategoryDto.alea_label,
               },
               selected: false,
-              });
-            }
+            }]
           })
-          this.categories = aleasByCategory;
-          this.updateComponent();
+        }
+        else {
+          aleasByCategory.find((cat) => (cat as AleaCategoryVM).category == aleaCategoryDto.category_name)?.aleas.push({
+            alea: {
+              id: aleaCategoryDto.alea_id,
+              name: aleaCategoryDto.alea_name,
+              label: aleaCategoryDto.alea_label,
+            },
+            selected: false,
+          });
+        }
       })
+      this.categories = aleasByCategory;
+      this.updateComponent();
+    })
+  }
+
+  /*********************TEAM MEMBERS*************************/
+
+  /**
+   * Checked property of input alert mail checkbox
+   * @param id 
+   * @returns 
+   */
+  isAlertMailChecked(id: number): boolean {
+    return this.selectedMailIds.includes(id);
+  }
+
+  /**
+   * Select team mates to alert
+   * @param event 
+   */
+  selectMailUser(event) {
+    const mailId = parseInt(event.target.value);
+    if (event.target.checked) {
+      this.selectedMailIds.push(mailId);
+    }
+    else {
+      let i = 0;
+      this.selectedMailIds.forEach((id: number) => {
+        if (id === mailId) {
+          this.selectedMailIds.splice(i, 1);
+          return;
+        }
+
+        i++;
+      });
+    }
+    this.alert.mailAlerts = this.mailAlerts.filter(x => this.selectedMailIds.some((n) => n === x.id))
+  }
+
+  /**
+   * Add another team mate
+   */
+  addTeamMember() {
+    this.mailAlertModal?.open();
+  }
+
+  /**
+   * Refresh team members after adding one from modal
+   */
+  getTeamMembersMail() {
+    this.alertApiService.getMailAlerts().subscribe((ma) => {
+      this.mailAlerts = ma;
+      this.updateComponent();
+    })
+  }
+
+  /**
+   * Update view
+   */
+  updateComponent() {
+    this.cd.markForCheck();
+  }
+
+  createAlert() {
+
+    if (this.formGroup.value.name != null) {
+      this.alert.name = this.formGroup.value.name;
+    }
+    if (this.alert.name == '' || this.alert.name == null) {
+      this.toastrService.error('Alerte incomplète', 'Vous devez donner un nom à votre alerte');
+      return;
+    }
+    else if (this.alert.aleas.length === 0) {
+      this.toastrService.error('Alerte incomplète', 'Vous devez selectionner au moins un type d\'aléa à surveiller');
+      return;
+    }
+    else if (this.alert.mailAlerts.length === 0) {
+      this.toastrService.error('Alerte incomplète', 'Vous devez selectionner au moins une personne à qui envoyer l\'adresse');
+      return;
     }
 
-    /*********************TEAM MEMBERS*************************/
+    this.endAlertModal?.open();
+    this.endAlertModal?.createAlert();
 
-    /**
-     * Checked property of input alert mail checkbox
-     * @param id 
-     * @returns 
-     */
-    isAlertMailChecked(id: number): boolean {
-      return this.selectedMailIds.includes(id);
-    }
-    
-    /**
-     * Select team mates to alert
-     * @param event 
-     */
-    selectMailUser(event){
-      const mailId = parseInt(event.target.value);
-      if(event.target.checked){
-        this.selectedMailIds.push(mailId);
-      }
-      else{
-        let i = 0;
-        this.selectedMailIds.forEach((id: number) => {
-          if(id === mailId) {
-            this.selectedMailIds.splice(i,1);
-            return;
-          }
-    
-          i++;
-        });
-      }
-      this.alert.mailAlerts = this.mailAlerts.filter(x => this.selectedMailIds.some((n) => n === x.id))
-    }
+    // prévenir l'utilisateur sil n'a pas selectionné de zone que la
+    // zone applicable est le monde entier et que cela peut le spammer (au moins s'il a selectionné les séismes)
 
-    /**
-     * Add another team mate
-     */
-    addTeamMember(){
-      this.mailAlertModal?.open();
-    }
-
-    /**
-     * Refresh team members after adding one from modal
-     */
-    getTeamMembersMail(){
-      this.alertApiService.getMailAlerts().subscribe((ma) => {
-        this.mailAlerts = ma;
-        this.updateComponent();
-      })
-    }
-
-    /**
-     * Update view
-     */
-    updateComponent(){
-      this.cd.markForCheck();
-    }
-
-    createAlert(){
-
-      if(this.formGroup.value.name != null){
-        this.alert.name = this.formGroup.value.name;
-      }
-      if(this.alert.name == '' || this.alert.name == null){
-        this.toastrService.error('Alerte incomplète','Vous devez donner un nom à votre alerte');
-        return;
-      }
-      else if(this.alert.aleas.length === 0){
-        this.toastrService.error('Alerte incomplète','Vous devez selectionner au moins un type d\'aléa à surveiller');
-        return;
-      }
-      else if(this.alert.mailAlerts.length === 0){
-        this.toastrService.error('Alerte incomplète','Vous devez selectionner au moins une personne à qui envoyer l\'adresse');
-        return;
-      }
-
-      this.endAlertModal?.open();
-      this.endAlertModal?.createAlert();
-
-      // prévenir l'utilisateur sil n'a pas selectionné de zone que la
-      // zone applicable est le monde entier et que cela peut le spammer (au moins s'il a selectionné les séismes)
-
-    }
+  }
 }
