@@ -5,7 +5,8 @@ import { environment } from 'src/environments/environment';
 @Component({
     selector: 'app-footer',
     templateUrl: './Footer.component.html',
-    styleUrls: ['./Footer.component.css']
+    styleUrls: ['./Footer.component.css'],
+    standalone: false
 })
 export class Footer implements OnInit {
 
