@@ -1,12 +1,10 @@
-import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
-
-import { AppModule } from './app/app.module';
 import { environment } from './environments/environment';
+import { bootstrapApplication } from '@angular/platform-browser';
+import { App } from './app/app.component';
+import { appConfig } from './app/app.config';
 
-
-platformBrowserDynamic().bootstrapModule(AppModule)
-  .catch(err => console.error(err));
-
-  if(environment.production){
-    window.console.log = () => {};
-  }
+if(environment.production){
+    // eslint-disable-next-line @typescript-eslint/no-empty-function
+    window.console.log = () => { }
+}
+bootstrapApplication(App, appConfig).catch((err) => console.error(err));
